Hooked up the legacy menu to the new mode lookup
