+++ resources/lib/constants.py
@@ -3,8 +3,7 @@ LIVE_EVENTS_MODE = 'LIVE_EVENTS'
 PLAY_MODE = 'PLAY'
 PLAY_ITEM_MODE = 'PLAY_ITEM'
 PLAY_TV_MODE = 'PLAY_TV'
-LIST_SPORTS_MODE = 'LIST_SPORTS'
-INDEX_SPORTS_MODE = 'INDEX_SPORTS'
+
 UPCOMING_MODE = 'UPCOMING'
 AUTHENTICATE_MODE = 'AUTHENTICATE'
 AUTHENTICATION_DETAILS_MODE = 'AUTHENTICATION_DETAILS'
